+++ axelor-production/src/main/java/com/axelor/apps/production/exceptions/IExceptionMessage.java
@@ -129,4 +129,12 @@ public interface IExceptionMessage {
       "A cancel reason must be selected" /*)*/;
 
   static final String MANUF_STOCK_MOVE_ERROR_1 = /*$$(*/ "All products has been consumed" /*)*/;
+
+  static final String UNIT_COST_CALCULATION_IMPORT_FAIL_ERROR = /*$$(*/ "Data import failed" /*)*/;
+
+  static final String UNIT_COST_CALCULATION_IMPORT_CSV_ERROR = /*$$(*/
+      "Uploaded file is not a CSV file" /*)*/;
+
+  static final String UNIT_COST_CALCULATION_NO_PRODUCT = /*$$(*/
+      "Please select an element (a product, a product category or a product family) to run calculation" /*)*/;
 }